[Progress_8]バリデーション、フラッシュメッセージの実装

+++ app/controllers/books_controller.rb
@@ -5,34 +5,59 @@ class BooksController < ApplicationController
   
   def create
     @book = Book.new(book_params)
+    @books = Book.all
     @book.user_id = current_user.id
-    @book.save
-    redirect_to books_path
+    @user = current_user
+    if @book.save
+      flash[:notice] = "You have created book successfully."
+      redirect_to book_path(@book.id)
+    else
+      render :index
+    end
   end
 
   def index
     @book = Book.new
     @books = Book.all
+    @user = current_user
   end
 
   def show
+    @book_new = Book.new
     @book = Book.find(params[:id])
+    @user = current_user
   end
   
   def edit
+    book = Book.find(params[:id])
+    unless book.user.id == current_user.id
+      redirect_to books_path
+    end
     @book = Book.find(params[:id])
   end
   
   def update
     book = Book.find(params[:id])
-    book.update(book_params)
-    redirect_to book_path(book.id)
+    unless book.user.id == current_user.id
+      redirect_to books_path
+    end
+    
+    @book = Book.find(params[:id])
+    if @book.update(book_params)
+      flash[:notice] = "You have updated book successfully."
+      redirect_to book_path(book.id)
+    else
+      render :edit
+    end
   end
   
   def destroy
     book = Book.find(params[:id])
+    unless book.user.id == current_user.id
+      redirect_to books_path
+    end
     book.destroy
-    redirect_to books_path
+    redirect_to user_path(current_user.id)
   end
   
   private
